#439 DrawTool Shape Tab Filter
