+++ src/external/Dropy/dropy.css
@@ -171,3 +171,6 @@ dd {
 .dropy.openUp .dropy__content > ul {
     transform: translateY(-100%);
 }
+.dropy.openHorizontal .dropy__content > ul {
+    display: flex;
+}
